backend/requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
cffi==1.14.2
colorama==0.4.3
cryptography==3.1
Django==3.1
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.9
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
